Feature/ON4T-120_ProposalService_refactoring - updated error message of updateProposal of the dbClient
